fold assembler into yaml loader
The assembler was just to tiny static functions used exclusively
by the yaml loader. Not worth the extra module.
